# ballet/txn: small changes to improve parse performance